+++ vite.config.ts
@@ -1,8 +1,35 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import { viteStaticCopy } from 'vite-plugin-static-copy'
 
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [
+    vue(),
+    viteStaticCopy({
+      targets: [
+        {
+          src: 'node_modules/@ricky0123/vad-web/dist/vad.worklet.bundle.min.js',
+          dest: './'
+        },
+        {
+          src: 'node_modules/@ricky0123/vad-web/dist/silero_vad_v5.onnx',
+          dest: './'
+        },
+        {
+          src: 'node_modules/@ricky0123/vad-web/dist/silero_vad_legacy.onnx',
+          dest: './'
+        },
+        {
+          src: 'node_modules/onnxruntime-web/dist/*.wasm',
+          dest: './'
+        },
+        {
+          src: 'node_modules/onnxruntime-web/dist/*.mjs',
+          dest: './'
+        }
+      ]
+    })
+  ],
   build: {
     sourcemap: false,
     minify: 'esbuild',